Make sure id matches id in neo4j-create-cedar-admin.cypher
The keycloak id of the adin user's record and the id in neo4j profile
must match. This means that if

cedar-infra-neo4j/config/neo4j-create-cedar-admin.cypher

refers to

_id:"https://metadatacenter.org/users/ab2a9696-291f-4705-b5e6-6c262266c506"

keycloak-realm.CEDAR.development.2023-07-05.json must have:

    "id" : "ab2a9696-291f-4705-b5e6-6c262266c506"

as well, or vice versa: set the id from the keyloak config in the
neo4j-create-cedar-admin.cypher script.