Add optional /aroma serving via arp-dataverse
Serve static files for /aroma in case we want to access aroma via the
arp-dataverse domain. This might be the case when CEDAR and Dataverse
is installed on the same host. In this case the 80 and 443 is served by
nginx and so we cannot have a haproxy or other way to serve AROMA.

We could serve directly via Dataverse itself, but serving via nginx is
more flexible, because we can easily update aroma: we just need to update
the static files on the host and it will be available in nginx.

For this /usr/share/nginx/html/aroma must be mapped to the host's location
of the aroma build, e.g., /home/dataverse/describo-react/build.