Use explicit CEDAR_VERSION=2.7.7 for cedar-terminology-server
This is our workaround for the API change in BioPortal, which now returns
a list of URL-s instead of a string for the "publication" field
of the submissions.

https://github.com/metadatacenter/cedar-terminology-server/commit/3be991d95022c1687dacc78f6c965691fbed7862

We use the 2.7.7 version of the reminology server, which fixes this issue.
All oter components are all right with their current 2.6.47 version.